git-svn-id: https://svn.eduke32.com/eduke32@1033 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/namesdyn.c
@@ -32,8 +32,8 @@ struct dynitem
 {
     char *s;
     int *v;
-    int vstat;
-    int val;
+    short vstat;
+    short val;
 };
 struct HASH_table dynnamesH = {4096, NULL};
